move attachment methods to attachmenthelper
