+++ Test/Altinn.Correspondence.Tests/MalwareScanResultControllerTests.cs
@@ -1,11 +1,10 @@
 ﻿using System.Net.Http.Json;
 using System.Text.Json;
 using System.Text;
-using System.Net.Http.Headers;
 using Altinn.Correspondence.API.Models.Enums;
 using Altinn.Correspondence.API.Models;
-using Altinn.Correspondence.Tests.Factories;
 using Altinn.Correspondence.Application.Configuration;
+using Altinn.Correspondence.Tests.Helpers;
 
 namespace Altinn.Correspondence.Tests;
 
@@ -32,7 +31,7 @@ public class MalwareScanResultControllerTests : IClassFixture<CustomWebApplicati
     public async Task MalwareScanFoundNoThreat_Success()
     {
         // Initialize
-        var attachmentId = await AttachmentFactory.GetPublishedAttachment(_senderClient, _responseSerializerOptions);
+        var attachmentId = await AttachmentHelper.GetPublishedAttachment(_senderClient, _responseSerializerOptions);
         var jsonBody = GetMalwareScanResultJson("Data/MalwareScanResult_NoThreatFound.json", attachmentId);
         await SendMalwareScanResult(jsonBody);
         // Get Scanned status
@@ -45,7 +44,7 @@ public class MalwareScanResultControllerTests : IClassFixture<CustomWebApplicati
     public async Task MalwareScanFoundMaliciousSignature_Success()
     {
         // Initialize
-        var attachmentId = await AttachmentFactory.GetPublishedAttachment(_senderClient, _responseSerializerOptions);
+        var attachmentId = await AttachmentHelper.GetPublishedAttachment(_senderClient, _responseSerializerOptions);
         var jsonBody = GetMalwareScanResultJson("Data/MalwareScanResult_Malicious.json", attachmentId);
         await SendMalwareScanResult(jsonBody);
